requirements.txt
flask==2.0.1
google-cloud-documentai==2.20.1
google-cloud-storage==2.10.0
python-dotenv==1.0.0
google-generativeai==0.3.1